did some linting fixes
