--- src/components/Profile/index.styles.ts
@@ -1,31 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    paddingTop: 100,
-    paddingHorizontal: 30,
-  },
-  loginContainer: {
-    alignItems: 'center',
-  },
-  textInput: {
-    fontSize: 32,
-    marginTop: 20,
-    width: '80%',
-  },
-  btnText: {
-    fontSize: 32,
-    color: 'steelblue',
-    marginTop: 50,
-  },
-  regTextContainer: {
-    alignItems: 'flex-end',
-    textAlign: 'right',
-  },
-  regText: {
-    fontSize: 18,
-    marginTop: 50,
-  },
-});
-
-export default styles;
