+++ src/screens/ProfileScreen/index.styles.ts
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    flex: 1,
+  },
+  logoutBtnContainer: {
+    alignItems: 'flex-end',
+  },
+  logoutText: {
+    fontSize: 18,
+    marginTop: 150,
+    color: 'steelblue',
+  },
+});
+
+export default styles;
