+++ src/screens/SplashScreen/index.styles.ts
@@ -1,0 +1,11 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    flex: 1,
+  },
+});
+
+export default styles;
