+++ src/screens/SplashScreen/index.tsx
@@ -6,6 +6,7 @@ import {ActivityIndicator, Platform, SafeAreaView, Text} from 'react-native';
 import {RootStackNavigation} from '../../navigation';
 import {StackActions} from '@react-navigation/native';
 import analytics from '@react-native-firebase/analytics';
+import styles from './index.styles';
 
 const SplashScreen = (): ReactElement => {
   const navigation = useNavigation<RootStackNavigation>();
@@ -19,11 +20,10 @@ const SplashScreen = (): ReactElement => {
         navigation.dispatch(StackActions.replace('LoginScreen'));
       }
     }, 1000);
-  }, []);
+  }, [navigation]);
 
   return (
-    <SafeAreaView
-      style={{alignItems: 'center', justifyContent: 'center', flex: 1}}>
+    <SafeAreaView style={styles.container}>
       <Text>Welcome to IW Training App</Text>
       <ActivityIndicator size={'large'} />
     </SafeAreaView>
